Improved description ripping for deviantART
Also, added some functions to AbstractHTMLRipper that should allow
description ripping to be added to other rippers as well.
hasDescriptionSupport() is a function that will be overridden by a
ripper that supports descriptions, and will trigger the description
ripper. getDescription will grab the description from a page, and must
be overridden if you want to grab a description with a ripper.