Added line breaks to deviantART ripper
Finally added the support for newlines that I've been going on about in
the other commits. Also got rid of a comment that shouldn't have been
there, as well as fixing the fact that I just broke everything two
commits back and didn't notice until just now. Sorry about that!