Improved handling of headless mode & OS-specific config
This adds changes the check for command-lines arguments to avoir throwing an error when no argument is passed on an headless system. It will instead gracefully show the help page to the user.
It also adds a check for the Mac OS-specific properties, to avoid setting them on other systems.
The OS-specific properties and the app initialization have been moved to after the Headless check for cleanliness (No need to set GUI properties on an headless app and no need to initialize if we're only going to show an help page).