Improve logging and exception handling
This starts the clean up of the exceptions around the code. It tries to be as specific as possible in the exceptions caught, to avoid simply throwing Exception itself.
It also start making a few improvements to the logging (Both the logger itself and various logged messages related to exceptions).