Make NfsfwRipper inherit from AbstractHTMLRipper
Also make it have queue support since there are some galleries that have only
subalbum links. In case of galleries with both images and subalbums, such as
the one in the unit test, first rip the images in the base album and then the
images in the subalbum (and save them to a subdirectory of the current album
directory).

Note that Nfsfw.com showed some read time-outs when ripping the same albums that
it ripped Ok before. Adding a timeout in getNextPage() seems to help somewhat,
but if any issues are encountered for now the simplest fix seems to be to wait
a while before trying to rip the album again.