Resolve xlecx error
xlecx has changed its URL from 'xlecx.com' to 'xlecx.org'. This change modifies the ripper.java to reflect that change, and to add the required https prefix. If the https prefix is not added, the image download fails with 'no protocol'. That seems a little janky, but it was used in #120.