RedditRipper: added gallery support
A gallery is a special kind of reddit post, which can have up to 20
images. They can be recognized by the is_gallery field in the JSON.
The gallery_data array describes the structure of the gallery by
referencing the image ids, it also contains the captions and links, which
are ignored by this implementation. The s field of each entry in
media_metadata array appears to reference the highest quality image.
getJsonURL and getGID were modified to handle reddit.com/gallery/id
links. parseJsonChild was modified to handle the gallery case. The new
method handleGallery parses gallery_data and media_metadata and adds the
images to download. A test for a gallery was added.