not build on tags, release already built in future
currently 2 builds happen, one for the main branch, one for the tag.
this should not be necessary, one build is sufficient, producing a
verison including the git hash. this can be converted to a release
later on, and ripme.json updated to this build.

the release procedure is thus: first commit, and let it build,
second, create a tag like '2.0.4' and mark the release as official,
then push an updated ripme.json with the version 2.0.4-12-487e38cc
to trigger clients prompting to updating ifself when it is run.