permit to override the jgitver version
for packaging a tar ball, jgitver takes the default version of 0.0.0,
not practical. putting a version into a file would be possible, but
we try to produce a new version with every commit, which then can be
released if tests pass, without rebuild. this means the source code
should NOT have the version in the commit, but derived.

for tar balls, we know the name of the tar, so its easy to set the
same version in a script, or put a desired version. pay attention,
that ripme checks the version towards github, and suggests an update,
if the running version is less than the released one.

fixes #145.