Update README.md: remove donation links, update maintainer list, announce merge of ripmeapp2 back into ripmeapp (#2050)
Hi! I was the primary maintainer of ripmeapp/ripme from about 2016-2018. I'm pleased to see that while that repo was abandoned, a fork of the project continued over here. I'm back to poke around a bit and wanted to sync up the repo as well as make it clear who the primary developer is on the project now. I've also made an announcement post on the subreddit to point to the new active development repo.

I've pushed the main of ripmeapp2/ripme to ripmeapp and I want to make this change as well to both repos to make it clear where active development is happening and who is the current primary developer.

I'm not guaranteeing I'll stick around but I might poke around from time to time. I definitely don't have the time to resume any kind of primary development here. I'm also not necessarily encouraging the resurrection of ripmeapp/ripme over this fork, but if @soloturn wants to accept the maintainer invitation to that repo and merge the forks back together, I'd do what I can to support that. I have a personal (non-work) software-dev-capable computer again so working on a project like this is back on the table.

This change also fixes a Markdown linter style complaint: bulleted lists are preferred to start lines with `-` to avoid ambiguity with `*` used for bold/italics.